+++ routeros/provider_schema_helpers.go
@@ -738,7 +738,7 @@ var (
 
 		// #447 routeros_ip_dhcp_server_config.store_leases_disk == "immediately"
 		// routeros_ipv6_nd_prefix.preferred_lifetime == "infinity"
-		if slices.Contains(timeControlWords, old) || slices.Contains(timeControlWords, old) {
+		if slices.Contains(timeControlWords, old) || slices.Contains(timeControlWords, new) {
 			return old == new
 		}
 
@@ -865,6 +865,45 @@ var (
 
 		return true
 	}
+
+	BytesEqual = func(k, old, new string, d *schema.ResourceData) bool {
+		if old == "" {
+			return false
+		}
+
+		if AlwaysPresentNotUserProvided(k, old, new, d) {
+			return true
+		}
+
+		if slices.Contains([]string{"unlimited"}, old) || slices.Contains([]string{"unlimited"}, new) {
+			return old == new
+		}
+
+		// Compare values 64M/64M <> 67108864/67108864 or 64M <> 67108864:
+		oldSet := strings.FieldsFunc(old, split)
+		newSet := strings.FieldsFunc(new, split)
+		if len(oldSet) != len(newSet) {
+			return false
+		}
+
+		for i, _ := range oldSet {
+			o, err := ParseByteValues(oldSet[i])
+			if err != nil {
+				panic("[BytesEqual] parse 'old' value error: " + err.Error())
+			}
+
+			n, err := ParseByteValues(newSet[i])
+			if err != nil {
+				panic("[BytesEqual] parse 'new' value error: " + err.Error())
+			}
+
+			if o != n {
+				return false
+			}
+		}
+
+		return true
+	}
 )
 
 // ImplicitSingleHostCIDR is a DiffSuppressFunc function that prevents a configuration such as `192.168.1.2/32`
@@ -878,7 +917,7 @@ func ImplicitSingleHostCIDR6(k, old, new string, d *schema.ResourceData) bool {
 	return new == old + "/128"
 }
 
-func buildReadFilter(m map[string]interface{}) []string {
+func buildReadFilter(m map[string]any) []string {
 	var res []string
 
 	for fieldName, fieldValue := range m {