+++ routeros/resource_container_config.go
@@ -9,6 +9,7 @@ func ResourceContainerConfig() *schema.Resource {
 	resSchema := map[string]*schema.Schema{
 		MetaResourcePath: PropResourcePath("/container/config"),
 		MetaId:           PropId(Name),
+		MetaSkipFields:   PropSkipFields("assumed_registry_url", "memory_current"),
 
 		"registry_url": {
 			Type:        schema.TypeString,
@@ -27,10 +28,10 @@ func ResourceContainerConfig() *schema.Resource {
 			Description: "Specifies the password for authentication (starting from ROS 7.8)",
 		},
 		"ram_high": {
-			Type:        schema.TypeString,
-			Optional:    true,
-			Default:     "0",
-			Description: "RAM usage limit. (0 for unlimited)",
+			Type:             schema.TypeString,
+			Optional:         true,
+			Description:      "RAM usage limit.",
+			DiffSuppressFunc: BytesEqual,
 		},
 		"tmpdir": {
 			Type:        schema.TypeString,